move cache utility apis from profile apis
