+++ spytest/spytest/main.py
@@ -15,6 +15,7 @@ from spytest.ftrace import print_ftrace, ftrace_reset
 from spytest.ftrace import ftrace_prefix
 
 import utilities.common as utils
+from utilities.cache import init_cache
 from utilities.profile import init_profile
 
 
@@ -116,7 +117,8 @@ def _parse_args(pre_parse=False):
 
     args, unknown = _parse_args_early(pre_parse)
 
-    # initialize profiling lib
+    # initialize cache and profiling libs
+    init_cache()
     init_profile()
 
     # parse the bucket options
@@ -198,7 +200,8 @@ def _parse_args(pre_parse=False):
         print_ftrace("setting environment {} = {}".format(name, value))
         os.environ[name] = value
 
-    # initialize profiling lib again
+    # initialize cache and profiling libs again
+    init_cache()
     init_profile()
 
     if args.exclude_devices:
